// types/chord-types.ts
export type ChordType = 
  | 'major' | 'minor' | 'dim' | 'augmented'
  | '7' | '9' | 'major7' | 'minor7'
  | 'major9' | 'minor9' | 'sus2' | 'sus4'
  | 'add9' | 'm7b5' | 'm11' | 'dim7'
  | 'major11' | 'major13' | '6' | '69'
  | 'minor6' | 'minor13' | 'minorMajor7'
  | '7sus4' | 'augmented7' | 'augmentedMajor7'
  | '11' | 'bass' | 'user' | '7b5'
  | '7#5' | '9sus' | '13sus' | '7sus'; 